ADG created by factory, continuing test suite

+++ contracts/libraries/DarwinSwapOracleLibrary.sol
@@ -21,7 +21,7 @@ library DarwinSwapOracleLibrary {
         price1Cumulative = IDarwinSwapPair(pair).price1CumulativeLast();
 
         // if time has elapsed since the last update on the pair, mock the accumulated price values
-        (uint112 reserve0, uint112 reserve1, uint32 blockTimestampLast) = IDarwinSwapPair(pair).getReserves();
+        (uint256 reserve0, uint256 reserve1, uint32 blockTimestampLast) = IDarwinSwapPair(pair).getReserves();
         if (blockTimestampLast != blockTimestamp) {
             // subtraction overflow is desired
             uint32 timeElapsed = blockTimestamp - blockTimestampLast;
